separate indexer from server
